Use snapml package (instead of pai4sk)
